# build/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "@infavy/api-server",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/channels/blocked": {
      "get": {
        "summary": "Get blocked channels for a user",
        "parameters": [
          {
            "in": "query",
            "name": "uid",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of blocked channels"
          }
        }
      }
    },
    "/api/v1/channels": {
      "get": {
        "summary": "List channels"
      },
      "post": {
        "summary": "Create a channel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/channels/{id}": {
      "get": {
        "summary": "Get channel by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "put": {
        "summary": "Update channel by id"
      },
      "delete": {
        "summary": "Delete channel by id"
      }
    },
    "/api/v1/channels/{id}/videos": {
      "get": {
        "summary": "Get videos for a channel",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
...
}